pyproject.toml
[project]
name = "typecats"
description = "Structure unstructured data for the purpose of static type checking"
authors = [{name = "Peter Gaultney", email = "pgaultney@xoi.io"}]
version = "2.1.0"
requires-python = ">=3.12,<3.15"
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
keywords = ["attrs", "cattrs", "typing", "serialization", "structuring"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "attrs >=25.4.0,<27.0.0",
    "cattrs >=26.1.0,<27.0.0",
]

[project.urls]
Homepage = "https://github.com/xoeye/typecats"
Repository = "https://github.com/xoeye/typecats"
Issues = "https://github.com/xoeye/typecats/issues"

[dependency-groups]
dev = [
    "black>=26.1.0,<27.0.0",
    "flake8>=7.3.0,<8.0.0",
    "isort>=8.0.1,<9.0.0",
    "mypy>=1.19.1,<2.0.0",
    "pylint>=4.0.5,<5.0.0",
    "pytest>=9.0.2,<10.0.0",
    "pytest-cov>=7.0.0,<8.0.0",
    "yamllint>=1.38.0,<2.0.0",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.pylint.basic]
good-names = ["i", "j", "k", "ex", "Run", "_", "Cat"]

[tool.pylint.format]
max-line-length = 120

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
